object representation
Objects are reference counted. The object header contains information
which allows a `Ptr<T>` to be cast to `Ptr<Any>` and back. It properly
handles allocation and deallocation through a `Context`, including
calling `T`'s drop even if the ptr is currently stored as `Ptr<Any>`.

Previously, the VM used `Ptr<Object>` and `Handle<T>` types to store
objects at runtime. Handles were the strongly typed variant of pointers.
All object types had to implement the ObjectType trait and then had to
be placed into the object `Repr` enum for the purpose of casting them at
runtime. This is now no longer necessary, and adding a new object type
only requires that the `Object` trait is implement for that type.