errors now go through `cx`
Made `Error::new` private and implemented `cx.error` which
is the only way to construct an error.

Added the `fail!` macro, which is equivalent to:
    return Err(cx.error(message, span))

Errors can now be constructed with a `None` span.
This is the same as constructing the error with an empty span.
The reason for this change is that you don't always have access
to a span without potentially significant costs. For example, the VM
only has access to spans through the instruction->span table, which
is stored in the VM state and not in the global `Context`. Instead of
attaching a span when the error is constructed, the span will be
attached when the error is returned from the dispatch loop. Care will
be taken to ensure the error path is cold, so that the cost is minimized
as much as possible.