start implementing various runtime object types
- `Ptr` now delegates `Eq`/`Ord`/`Hash` to the inner `T`
- `Ptr` also provides `ptr_eq` and `ptr_hash` methods for directly
  comparing pointers.
- Added rough sketch of:
  - `Function`
  - `Class`
  - `Module`
  - `List`
  - `Table`
  - `String`

Note that all mutability will be handled via `RefCell` inside the
runtime types themselves. Previously, it was handled via an `UnsafeCell`
in `Ptr`, but this made usage dangerous and really unsafe.
This time around, all references and pointers will assume everything is
immutable, and if a type (either builtin or user-defined) needs
mutability, it must explicitly provide it by wrapping the mutable fields
in `RefCell`.