begin work on instruction encoding - add `BytecodeBuilder` - add `ConstantPoolBuilder` - add macros and traits to define instructions and operands
The `instructions` macro generates a struct for each instruction,
and implements the `Instruction` trait for it. Most of the actual work
happens in the `Operand` trait implementations. This is different from
how it was handled in V1, where the macro did significantly more work.

The definition is not yet complete though, it still needs disassembly.