Implement jump instruction emit
See `docs/instructions.md` for how it works.
The short version is that we reserve a constant pool entry for each jump
offset, and commit it if the jump offset overflows during patching.
In practice, most jumps will have offsets that fit into 8 bits, so they
will the offset will be encoded directly using just 1 byte.