allow regalloc to reuse partial slices
Before, if the register file contained only 2 free registers at the top
and the emitter requested 4, it would have to allocate 4 completely
fresh registers.

Now the allocator has the ability to reuse those 2 top-most registers
and only allocate 2 fresh ones, which greatly improves the register
allocator's ability to reuse registers.

For example, consider the following code:
    f(0)
    f(1, 2)
Before this change, the second call cannot reuse the registers of the
first call, because the first call only used 2 registers. This means
that this function will need 5 registers total.
After this change, the second call can re-use the 2 registers used in
the first call, and only allocates 1 extra register.
In total, this function will only need 3 registers now.

This kind of change has compounding interest, and in the grand scheme
doesn't cost all that much, because the added time complexity
is only O(M) where M is the number of registers being allocated.